Refactor code and add views
