+++ csharp/Profiler/DurationView.cs
@@ -1,0 +1,8 @@
+﻿namespace Profiler;
+
+public class DurationView : LineProfileView
+{
+    public DurationView(LineProfile line) : base(line)
+    {
+    }
+}
