+++ csharp/Profiler/File.cs
@@ -1,11 +1,10 @@
 ﻿using System.Collections.Generic;
 
-namespace Profiler
+namespace Profiler;
+
+public class File
 {
-    public class File
-    {
-        public string Path { get; internal set; }
-        public string Name { get; internal set; }
-        public Dictionary<int, LineProfile> Lines { get; internal set; }
-    }
+    public string Path { get; internal set; }
+    public string Name { get; internal set; }
+    public Dictionary<int, LineProfile> Lines { get; internal set; }
 }
