+++ csharp/Profiler/FunctionDurationView.cs
@@ -1,0 +1,8 @@
+﻿namespace Profiler;
+
+public class FunctionDurationView : LineProfileView
+{
+    public FunctionDurationView(LineProfile line) : base(line)
+    {
+    }
+}
