+++ csharp/Profiler/FunctionHitCountView.cs
@@ -1,0 +1,8 @@
+﻿namespace Profiler;
+
+public class FunctionHitCountView : LineProfileView
+{
+    public FunctionHitCountView(LineProfile line) : base(line)
+    {
+    }
+}
