+++ csharp/Profiler/FunctionSelfDurationView.cs
@@ -1,0 +1,8 @@
+﻿namespace Profiler;
+
+public class FunctionSelfDurationView : LineProfileView
+{
+    public FunctionSelfDurationView(LineProfile line) : base(line)
+    {
+    }
+}
