+++ csharp/Profiler/HitCountView.cs
@@ -1,0 +1,8 @@
+﻿namespace Profiler;
+
+public class HitCountView : LineProfileView
+{
+    public HitCountView(LineProfile line) : base(line)
+    {
+    }
+}
