+++ csharp/Profiler/ITracer.cs
@@ -1,10 +1,9 @@
 ﻿using System.Management.Automation;
 using System.Management.Automation.Language;
 
-namespace Profiler
+namespace Profiler;
+
+public interface ITracer
 {
-    public interface ITracer
-    {
-        void Trace(IScriptExtent extent, ScriptBlock scriptBlock, int level, string functionName, string moduleName);
-    }
+    void Trace(IScriptExtent extent, ScriptBlock scriptBlock, int level, string functionName, string moduleName);
 }
