+++ csharp/Profiler/MemoryView.cs
@@ -1,0 +1,8 @@
+﻿namespace Profiler;
+
+public class MemoryView : LineProfileView
+{
+    public MemoryView(LineProfile line) : base(line)
+    {
+    }
+}
