+++ csharp/Profiler/SelfDurationView.cs
@@ -1,0 +1,8 @@
+﻿namespace Profiler;
+
+public class SelfDurationView : LineProfileView
+{
+    public SelfDurationView(LineProfile line) : base(line)
+    {
+    }
+}
