+++ csharp/Profiler/SelfMemoryView.cs
@@ -1,0 +1,8 @@
+﻿namespace Profiler;
+
+public class SelfMemoryView : LineProfileView
+{
+    public SelfMemoryView(LineProfile line) : base(line)
+    {
+    }
+}
