+++ csharp/Profiler/SpeedScope/SpeedScopeProfile.cs
@@ -1,14 +1,13 @@
 ﻿using System.Collections.Generic;
 
-namespace Profiler.SpeedScope
+namespace Profiler.SpeedScope;
+
+public class SpeedScopeProfile
 {
-    public class SpeedScopeProfile
-    {
-        public string Type { get; set; }
-        public string Name { get; set; }
-        public string Unit { get; set; }
-        public int StartValue { get; set; }
-        public double EndValue { get; set; }
-        public List<SpeedScopeEvent> Events { get; set; }
-    }
+    public string Type { get; set; }
+    public string Name { get; set; }
+    public string Unit { get; set; }
+    public int StartValue { get; set; }
+    public double EndValue { get; set; }
+    public List<SpeedScopeEvent> Events { get; set; }
 }
