+++ demo4-recursion.ps1
